BUG 1098526: Confirm column delete
QA:
* Make sure bug is fixed

+++ public/javascripts/widgets/dataset-grid.js
@@ -799,23 +799,33 @@
                 var cols = [];
                 var successCount = 0;
                 $.each(selCols, function(colId, val)
+                        { cols.push(colId); });
+                var multiCols = cols.length > 1;
+                if (confirm('Do you want to delete the ' +
+                    (multiCols ? cols.length + ' selected columns' :
+                        'selected column') + '? All data in ' +
+                    (multiCols ? 'these columns' : 'this column') +
+                    ' will be removed!'))
                 {
-                    cols.push(colId);
-                    $.ajax({url: '/views/' + view.id + '/columns/' +
-                        colId + '.json', type: 'DELETE',
-                        contentType: 'application/json',
-                        complete: function()
-                        {
-                            successCount++;
-                            if (successCount == cols.length)
+                    $.each(cols, function(i, colId)
+                    {
+                        $.ajax({url: '/views/' + view.id + '/columns/' +
+                            colId + '.json', type: 'DELETE',
+                            contentType: 'application/json',
+                            complete: function()
                             {
-                                model.deleteColumns(cols);
-                                $(document).trigger(blist.events.COLUMNS_CHANGED);
-                            }
-                        }});
-                });
-                // Hide columns so they disappear immediately
-                datasetObj.showHideColumns(cols, true, true);
+                                successCount++;
+                                if (successCount == cols.length)
+                                {
+                                    model.deleteColumns(cols);
+                                    $(document)
+                                        .trigger(blist.events.COLUMNS_CHANGED);
+                                }
+                            }});
+                    });
+                    // Hide columns so they disappear immediately
+                    datasetObj.showHideColumns(cols, true, true);
+                }
                 break;
         }
         // Update the grid header to reflect updated sorting, filtering
